Fix checks for gmtime_r and localtime_r availability.
Actually check for the functions in configure.ac and CMakeLists.txt,
which is the only way of definitively testing for them, since feature
test macros such as _POSIX_C_SOURCE are supposed to be used to tell
system headers what functions should be exposed, and the application
shouldn't check their values.

Add a warning for cases where gmtime() or localtime() are used.

Also fix Http_Cookies where gmtime() was still being used.